+++ include/databento/detail/tcp_client.hpp
@@ -7,21 +7,11 @@
 #include <string_view>
 
 #include "databento/detail/scoped_fd.hpp"  // ScopedFd
+#include "databento/ireadable.hpp"
 
 namespace databento::detail {
 class TcpClient {
  public:
-  enum class Status : std::uint8_t {
-    Ok,
-    Timeout,
-    Closed,
-  };
-
-  struct Result {
-    std::size_t read_size;
-    Status status;
-  };
-
   struct RetryConf {
     std::uint32_t max_attempts{1};
     std::chrono::seconds max_wait{std::chrono::minutes{1}};
@@ -33,11 +23,11 @@ class TcpClient {
   void WriteAll(std::string_view str);
   void WriteAll(const std::byte* buffer, std::size_t size);
   void ReadExact(std::byte* buffer, std::size_t size);
-  Result ReadSome(std::byte* buffer, std::size_t max_size);
+  IReadable::Result ReadSome(std::byte* buffer, std::size_t max_size);
   // Passing a timeout of 0 will block until data is available of the socket is
   // closed, the same behavior as the Read overload without a timeout.
-  Result ReadSome(std::byte* buffer, std::size_t max_size,
-                  std::chrono::milliseconds timeout);
+  IReadable::Result ReadSome(std::byte* buffer, std::size_t max_size,
+                             std::chrono::milliseconds timeout);
   // Closes the socket.
   void Close();