+++ src/live.cpp
@@ -79,18 +79,26 @@ LiveBuilder& LiveBuilder::ExtendUserAgent(std::string extension) {
   return *this;
 }
 
+LiveBuilder& LiveBuilder::SetCompression(Compression compression) {
+  compression_ = compression;
+  return *this;
+}
+
 databento::LiveBlocking LiveBuilder::BuildBlocking() {
   Validate();
   if (gateway_.empty()) {
     return databento::LiveBlocking{log_receiver_,   key_,
                                    dataset_,        send_ts_out_,
                                    upgrade_policy_, heartbeat_interval_,
-                                   buffer_size_,    user_agent_ext_};
+                                   buffer_size_,    user_agent_ext_,
+                                   compression_};
   }
-  return databento::LiveBlocking{
-      log_receiver_, key_,           dataset_,        gateway_,
-      port_,         send_ts_out_,   upgrade_policy_, heartbeat_interval_,
-      buffer_size_,  user_agent_ext_};
+  return databento::LiveBlocking{log_receiver_,   key_,
+                                 dataset_,        gateway_,
+                                 port_,           send_ts_out_,
+                                 upgrade_policy_, heartbeat_interval_,
+                                 buffer_size_,    user_agent_ext_,
+                                 compression_};
 }
 
 databento::LiveThreaded LiveBuilder::BuildThreaded() {
@@ -99,12 +107,15 @@ databento::LiveThreaded LiveBuilder::BuildThreaded() {
     return databento::LiveThreaded{log_receiver_,   key_,
                                    dataset_,        send_ts_out_,
                                    upgrade_policy_, heartbeat_interval_,
-                                   buffer_size_,    user_agent_ext_};
+                                   buffer_size_,    user_agent_ext_,
+                                   compression_};
   }
-  return databento::LiveThreaded{
-      log_receiver_, key_,           dataset_,        gateway_,
-      port_,         send_ts_out_,   upgrade_policy_, heartbeat_interval_,
-      buffer_size_,  user_agent_ext_};
+  return databento::LiveThreaded{log_receiver_,   key_,
+                                 dataset_,        gateway_,
+                                 port_,           send_ts_out_,
+                                 upgrade_policy_, heartbeat_interval_,
+                                 buffer_size_,    user_agent_ext_,
+                                 compression_};
 }
 
 void LiveBuilder::Validate() {