+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.47.0 - 2026-02-04
+
+### Enhancements
+- Added Zstd compression support to live clients which can be enabled with
+`LiveBuilder::SetCompression()`. It's disabled by default
+- Added `Compression()` getter to `LiveBlocking` and `LiveThreaded`
+- Upgraded default `httplib` version to 0.30.1
+
+### Breaking changes
+- Added an overload to the `ReadSome` method on `IReadable` for timeout support
+
 ## 0.46.1 - 2026-01-27
 
 ### Enhancements